fix: try to minimize the change of missing a new lap time
Previously if two consecutive laps happened to have exactly the same times, then we would have missed a new lap. Compare the sector times too now. If somehow sector times are equal too, then we don't even need to update the lap.

Also only check them if we are expecting a new lap time.

As a note we need to check lap time and new lap separately because they may not be in complete sync.