fix: special case ACC's lap time update
This fixes a bug where in qualy/practice session joining mid session misses lap invalidation. Thus possibly setting invalid lap as best lap. The order remains correct (it's set by the position received from ACC) but an invalid lap is shown as a lap.  Since the order is supposed to be based on the best lap, this could show weird discrepancy between position and lap time.