feat: allow removing leaderboard from rotation if there is only single class or cup in session
In single class/cup sessions the overall and class/cup leaderboards are equivalent. Thus there is really no need to cycle through all of them if both are enabled.

This commit allows to set which leaderboard will be removed from rotation order if there is single class/cup in session.