feat: rework class information setting to add "same as" option to classes
This allows to register multiple classes as one without having to change the class of every single car.

For example in R3E multiple classes are grouped together in game too. DTM 2023 and GTR 3 are both GT3 cars and are shown as a single class.

Similarly in AMS2 GT3 and GT3_gen2 are both GT3 cars.