feat: replace rename class option with duplicate
Rename had various issues and it more or less behaved like duplicate except it also changed cars to use that new class. Changing classes on cars will be a new option in car setting tab. I think it's clearer to separate these two out.

This also adds a duplicated from field to class info. This is used to walk back and see what base a duplicated field should use. It allows for reset to reset back to the original base which is expected.