feat: remove our own ACC broadcasting connection
The reason is that I don't think it's actually necessary.
One could permanently run ACC in spectator mode from SimHub, if they need spectator mode.
When the user is driving all of the ACC data is still available.
Tested couple of dashboards and they seem to be all fine.

I'll add another branch with this removed code, if there actually are issues and we may need all of this removed code.